OkMakieToolkits

Documentation for OkMakieToolkits.

OkMakieToolkits.PairedCircleGenerator

OkMakieToolkits.TimeAsX

OkMakieToolkits.blankaxis!

OkMakieToolkits.circle_inner

OkMakieToolkits.circle_outer

OkMakieToolkits.dataoverview!

OkMakieToolkits.datetimeticks!

OkMakieToolkits.datetimeticks!

OkMakieToolkits.datetimeticks!

OkMakieToolkits.expandxylim!

OkMakieToolkits.expandylim!

OkMakieToolkits.getxylimits

OkMakieToolkits.mkgen_hollow

OkMakieToolkits.mkgen_solid

OkMakieToolkits.secondyaxis

OkMakieToolkits.shrinkylim!

OkMakieToolkits.PairedCircleGenerator — Type

PairedCircleGenerator(outer_radius, linewidth) returns a pair of functions, solid and hollow, which generates solid and hollow circle of the same outer radius.

Example

PCG = PairedCircleGenerator(80, 1)
mksz = [1,2,3]
mksz2 = [0.8,1.7,2.5]
scatter!(...; markersize = mksz, marker = PCG.solid.(mksz))
scatter!(...; markersize = mksz2, marker = PCG.hollow.(mksz2))

OkMakieToolkits.TimeAsX — Type

TimeAsX bind Integer indices to a DateTime array of the same size.

OkMakieToolkits.blankaxis! — Method

blankaxis!(fpos, xmin::T, ymin::T, xmax::T, ymax::T) where {T<:AbstractFloat} creates an empty Makie Axis object at grid position fpos.

Example

f = Figure(;resolution=(1400,1000))
gleft = f[1:3, 1] = GridLayout()
gright = f[1:3, 2:3] = GridLayout()
fpos = gright[0, :]
blankaxis!(fpos, xmin, ymin, xmax, ymax)

OkMakieToolkits.circle_inner — Method

circle_inner(outer_radius,linewidth,markersize) creates the inner circle as a vector of points, where the linewidth is scaled by markersize.

OkMakieToolkits.circle_outer — Method

circle_outer(outer_radius) creates a circle as a vector of points.

OkMakieToolkits.dataoverview! — Method

dataoverview!(df::DataFrame, xtype::DataType; gridsize = gridsize, set_left_edge=leftfn, set_right_edge=rightfn) plot the overview of data in DataFrame df. xtype specify the only column type (e.g., DateTime) for x-axis.

Example

using Dates, DataFrames, OkMakieToolkits, Makie
a = randn(100)
a[1:20] .= NaN
b = Vector{Union{Missing, Float64}}(undef, 100)
b[1:70] .= randn(70)
b[71:90] .= NaN
b[91:100] .= missing
table_nan = DataFrame(:a => [1,2,NaN], :b => [missing,missing,5], :dt => collect(DateTime("2022-01-01T00:00:00"):Day(1):DateTime("2022-01-03T00:00:00")))
f, ax, dfxx = dataoverview!(table_nan, DateTime; resolution = (800,500))
Makie.save("Fig_dataoverview.eps", f)

OkMakieToolkits.datetimeticks! — Method

datetimeticks!(ax2, DTX::TimeAsX; kwargs...). See TimeAsX

OkMakieToolkits.datetimeticks! — Method

Similar to datetimeticks!(ax2, t::Vector{DateTime}, x_a::Vector), datetimeticks!(ax2, t::Vector{DateTime}, x_a::Vector, tinc::DatePeriod; datestrformat = "yyyy/mm/dd", modify_fn = identity) return xticks, xticklabels with xticks forced to increase with step tinc, and be modify by function modify_fn.

Example

datetimeticks!(
    ax2,
    [DateTime(2012,2,5,3,15,0), DateTime(2012,3,5,3,15,0)],
    [0, 1], # `x` for `Makie.plot(x, ...)`
    Day(3); # tick every 3 days
    datestrformat = "yyyy/mm/dd",
    modify_fn = x -> floor.(x, Day)
)

OkMakieToolkits.datetimeticks! — Method

datetimeticks!(ax2, t::Vector{DateTime}, x::Vector; datestrformat = "yyyy/mm/dd") set x ticks to datestr format. t is the DateTime array that is not supported by Makie, x is a arbitrarily defined series of numbers that corresponds to t for Makie.plot. x and t must be the same length and should be pairwisely mapped.

Example:

t = df.datetime
x = Dates.datetime2epochms.(t)
x1 = x .- x[1] # to avoid glitchs in plotting with CairoMakie (due to the too-large values)
y = df.soil_water
CairoMakie.scatter!(ax, x, y, markersize =3)
datetimeticks!(ax2, t,x)

OkMakieToolkits.expandxylim! — Method

expandxylim!(ax, upper_expand_x, lower_expand_x, lower_expand_y, upper_expand_y) TODO: merge with expandylim! in makietools.jl

OkMakieToolkits.expandylim! — Method

expandylim!(ax, upper_expand) expand the upper limit of y axis to a ratio of upper_expand of the original y range (ymax - ymin).

Example

expandylim!(ax, 0.25)

OkMakieToolkits.getxylimits — Method

Example

xmin, ymin, xmax, ymax = getxylimits(ax)

OkMakieToolkits.mkgen_hollow — Method

mkgen_hollow(markersize::Number; linewidth=1, outer_radius=80) generates hollow circle as a Polygon, which can be used for the marker kwarg of scatter!.

OkMakieToolkits.mkgen_solid — Method

mkgen_solid(;outer_radius=80) generates solid circle as a Polygon, which can be used for the marker kwarg of scatter!.

OkMakieToolkits.secondyaxis — Method

secondyaxis(f_grid; color = :black, ylabel = "y") returns an Axis object of y axis to the right-hand side, with default settings for yyplot.

Example

f = Figure()
ax_right = secondyaxis(f[1,1]; color=:red)

that you can do lines!(ax_right, ...)

Further reading

Is it possible to add a secondary y axis?

OkMakieToolkits.shrinkylim! — Method

shrinkylim!(ax, ydatas::Vector{<:Number}; extent = 0.15) fit y-limits to the input ydatas to with a certain extent.