OkMakieToolkits

Documentation for OkMakieToolkits.

OkMakieToolkits.PairedCircleGenerator

OkMakieToolkits.PredictData

OkMakieToolkits.PredictData

OkMakieToolkits.RangedTimeAsX

OkMakieToolkits.RangedTimeAsX

OkMakieToolkits.RangedTimeAsX

OkMakieToolkits.RangedTimeAsX

OkMakieToolkits.TimeAsX

OkMakieToolkits.TimeAsX

Makie.linkaxes!

OkMakieToolkits.approxid

OkMakieToolkits.blankaxis!

OkMakieToolkits.circle_inner

OkMakieToolkits.circle_outer

OkMakieToolkits.datetimeticks!

OkMakieToolkits.datetimeticks!

OkMakieToolkits.datetimeticks!

OkMakieToolkits.datetimeticks!

OkMakieToolkits.expandxylim!

OkMakieToolkits.expandylim!

OkMakieToolkits.get_plot_elements

OkMakieToolkits.getlimits

OkMakieToolkits.getxylimits

OkMakieToolkits.mkgen_hollow

OkMakieToolkits.mkgen_solid

OkMakieToolkits.secondyaxis

OkMakieToolkits.shrinkylim!

OkMakieToolkits.timeseriesplot

OkMakieToolkits.timeseriesplot!

OkMakieToolkits.PairedCircleGenerator — Type

PairedCircleGenerator(outer_radius, linewidth) returns a pair of functions, solid and hollow, which generates solid and hollow circle of the same outer radius.

Example

PCG = PairedCircleGenerator(80, 1)
mksz = [1,2,3]
mksz2 = [0.8,1.7,2.5]
scatter!(...; markersize = mksz, marker = PCG.solid.(mksz))
scatter!(...; markersize = mksz2, marker = PCG.hollow.(mksz2))

OkMakieToolkits.PredictData — Type

struct PredictData{T<:Real}
    pred::T
    real::T
    fold::Int
end

OkMakieToolkits.PredictData — Method

An object that binds real data and predictions.

PredictData(pred, real) = PredictData(pred, real, 0)

OkMakieToolkits.RangedTimeAsX — Type

struct RangedTimeAsX x::AbstractRange t::StepRange dx::Int dt::TimePeriod end

Similar as TimeAsX, but RangedTimeAsX guarantee the step size to be uniform.

OkMakieToolkits.RangedTimeAsX — Method

RangedTimeAsX(x::AbstractRange, t::StepRange). Error will occur if x, t are not of the same length.

OkMakieToolkits.RangedTimeAsX — Method

RangedTimeAsX(TX::TimeAsX) convert a TimeAsX object to RangedTimeAsX if possible.

OkMakieToolkits.RangedTimeAsX — Method

RangedTimeAsX(x::Vector{<:Int}, t::Vector{<:DateTime}).

Error will occur if x and t can not be converted to ranges of uniform step.

OkMakieToolkits.TimeAsX — Type

TimeAsX(x, t; check_approxid = false) bind Integer indices to a DateTime array of the same size.

Example

t = DateTime(2018,1,1):Day(1):DateTime(2018,1,10)
x = 5:14
TX = TimeAsX(x, t)
TX[1]

TimeAsX([5], [DateTime("2018-01-01T00:00:00")])

If check_approxid = true, it checks whether dx/dt is approximately identical anywhere.

OkMakieToolkits.TimeAsX — Method

Example

t = DateTime(2018,1,1):Day(1):DateTime(2018,1,10)
TX = TimeAsX(t)

Makie.linkaxes! — Method

Given a figure, link all axes in f.content ([... for ax in f.content if ax isa Axis]).

Inspired by this reply

OkMakieToolkits.approxid — Method

approxid(a, b) returns a if a and b are approximately identical. It passes keyword arguments to isapprox.

OkMakieToolkits.blankaxis! — Method

blankaxis!(fpos, xmin::T, ymin::T, xmax::T, ymax::T) where {T<:AbstractFloat} creates an empty Makie Axis object at grid position fpos.

Example

f = Figure(;resolution=(1400,1000))
gleft = f[1:3, 1] = GridLayout()
gright = f[1:3, 2:3] = GridLayout()
fpos = gright[0, :]
blankaxis!(fpos, xmin, ymin, xmax, ymax)

OkMakieToolkits.circle_inner — Method

circle_inner(outer_radius,linewidth,markersize) creates the inner circle as a vector of points, where the linewidth is scaled by markersize.

OkMakieToolkits.circle_outer — Method

circle_outer(outer_radius) creates a circle as a vector of points.

OkMakieToolkits.datetimeticks! — Method

datetimeticks!(ax2, DTX::TimeAsX, tinc::DatePeriod; kwargs...). See TimeAsX

OkMakieToolkits.datetimeticks! — Method

datetimeticks!(ax2, DTX::TimeAsX; kwargs...). See TimeAsX

OkMakieToolkits.datetimeticks! — Method

Similar to datetimeticks!(ax2, t::Vector{DateTime}, x_a::Vector), datetimeticks!(ax2, t::Vector{DateTime}, x_a::Vector, tinc::DatePeriod; datestrformat = "yyyy/mm/dd", modify_fn = identity) return xticks, xticklabels with xticks forced to increase with step tinc, and be modify by function modify_fn.

Example

datetimeticks!(
    ax2,
    [DateTime(2012,2,5,3,15,0), DateTime(2012,3,5,3,15,0)],
    [0, 1], # `x` for `Makie.plot(x, ...)`
    Day(3); # tick every 3 days
    datestrformat = "yyyy/mm/dd",
    modify_fn = x -> floor.(x, Day)
)

OkMakieToolkits.datetimeticks! — Method

datetimeticks!(ax2, t::Vector{DateTime}, x::Vector; datestrformat = "yyyy/mm/dd") set x ticks to datestr format. t is the DateTime array that is not supported by Makie, x is a arbitrarily defined series of numbers that corresponds to t for Makie.plot. x and t must be the same length and should be pairwisely mapped.

Example:

t = df.datetime
x = Dates.datetime2epochms.(t)
x1 = x .- x[1] # to avoid glitchs in plotting with CairoMakie (due to the too-large values)
y = df.soil_water
CairoMakie.scatter!(ax, x, y, markersize =3)
datetimeticks!(ax2, t,x)

OkMakieToolkits.expandxylim! — Method

expandxylim!(ax, upper_expand_x, lower_expand_x, lower_expand_y, upper_expand_y) TODO: merge with expandylim! in makietools.jl

OkMakieToolkits.expandylim! — Method

expandylim!(ax, upper_expand) expand the upper limit of y axis to a ratio of upper_expand of the original y range (ymax - ymin).

Example

expandylim!(ax, 0.25)

OkMakieToolkits.get_plot_elements — Method

Referencing: https://docs.makie.org/stable/examples/blocks/legend/index.html#creatinglegendentries_manually

OkMakieToolkits.getlimits — Method

getlimits(ax::Axis, dim) returns the finallimits[] of dimension dim.

Example

(x0, x1) = convert.(Float64, getlimits(ax::Axis, 1))
(y0, y1) = convert.(Float64, getlimits(ax::Axis, 2))

OkMakieToolkits.getxylimits — Method

Example

xmin, ymin, xmax, ymax = getxylimits(ax)

OkMakieToolkits.mkgen_hollow — Method

mkgen_hollow(markersize::Number; linewidth=1, outer_radius=80) generates hollow circle as a Polygon, which can be used for the marker kwarg of scatter!.

OkMakieToolkits.mkgen_solid — Method

mkgen_solid(;outer_radius=80) generates solid circle as a Polygon, which can be used for the marker kwarg of scatter!.

OkMakieToolkits.secondyaxis — Method

secondyaxis(f_grid; color = :black, ylabel = "y") returns an Axis object of y axis to the right-hand side, with default settings for yyplot.

Example

f = Figure()
ax_right = secondyaxis(f[1,1]; color=:red)

that you can do lines!(ax_right, ...)

Further reading

Is it possible to add a secondary y axis?

OkMakieToolkits.shrinkylim! — Method

shrinkylim!(ax, ydatas::Vector{<:Number}; extent = 0.15) fit y-limits to the input ydatas to with a certain extent.

OkMakieToolkits.timeseriesplot! — Method

This is an experimental plot function of timeseriesplot. It takes x::AbstractVector{<:Real} and values::AbstractVector{<:PredictData}, where

x and values automatically become Observable inside the plot! function due to Makie's magic.

additional Observables should be defined if you want the plot to be updated on the change of data.

In this recipe, we do not define color of lines/scatters as Observable that color won't update.

Example

f = Figure();
timeseriesplot(f[1,1], x, values; testcolor=:brown2, realcolor=:blue, markersize=3)
# or
ax = Axis(f[1,2])
timeseriesplot!(ax, x, values)
# or
(f, ax, p) = timeseriesplot(TX.x[(firsttrain+1):end], [OT...])

See also

This function is created referencing Example: Stock Chart

You may see The recipe of raincloud for a more advanced utility of recipe. Noted that they use @recipe(RainClouds, category_labels, data_array) do scene, not just @recipe(RainClouds) do scene

OkMakieToolkits.timeseriesplot — Method

This is an experimental plot function of timeseriesplot. It takes x::AbstractVector{<:Real} and values::AbstractVector{<:PredictData}, where

x and values automatically become Observable inside the plot! function due to Makie's magic.

additional Observables should be defined if you want the plot to be updated on the change of data.

In this recipe, we do not define color of lines/scatters as Observable that color won't update.

Example

f = Figure();
timeseriesplot(f[1,1], x, values; testcolor=:brown2, realcolor=:blue, markersize=3)
# or
ax = Axis(f[1,2])
timeseriesplot!(ax, x, values)
# or
(f, ax, p) = timeseriesplot(TX.x[(firsttrain+1):end], [OT...])

See also

This function is created referencing Example: Stock Chart

You may see The recipe of raincloud for a more advanced utility of recipe. Noted that they use @recipe(RainClouds, category_labels, data_array) do scene, not just @recipe(RainClouds) do scene